% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitm-helpers.R
\name{call_mitm}
\alias{call_mitm}
\title{Spawn a mitmdump background process and return the process id}
\usage{
call_mitm(args)
}
\arguments{
\item{args}{see \code{\link[sys:exec_background]{sys::exec_background()}}}
}
\description{
Spawn a mitmdump background process and return the process id
}
