% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-check.R
\name{check_check}
\alias{check_check}
\title{Check the HAR from the mitm_check() call}
\usage{
check_check(mitm_check_har)
}
\arguments{
\item{mitm_check_har}{HAR file from \code{\link[=mitm_check]{mitm_check()}}}
}
\description{
Check the HAR from the mitm_check() call
}
