% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_mitm}
\alias{install_mitm}
\title{Helper to get mitmproxy installed}
\usage{
install_mitm()
}
\description{
It doesn't so much as "install" anything as guide you on the path.
}
\seealso{
Other mitm_helpers: \code{\link{call_mitm}},
  \code{\link{mitm_help}}, \code{\link{start_mitm}}
}
