% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-cert.R
\name{install_mitm_certificate}
\alias{install_mitm_certificate}
\title{Helper to install mitmproxy root CA certificate}
\usage{
install_mitm_certificate()
}
\description{
This won't install the certificate in Firefox. You will need to do that manually.
}
