% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middlechild.R
\docType{package}
\name{middlechild}
\alias{middlechild}
\alias{middlechild-package}
\title{Tools to Intercept, Validate and Consume Web/Network Traffic}
\description{
The 'mitmproxy' \url{https://mitmproxy.org/} project provides tools
to intercept, modify and/or introspect network traffic. Methods are provided to
download, install, configure and launch 'mitmproxy' plus introspect and validate
network captures.
}
\author{
Ildiko Czeller (czeildi@gmail.com)

Karthik Ram (karthik.ram@gmail.com)

Bob Rudis (bob@rud.is)

Kara Woo
}
