% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitm-helpers.R
\name{mitm_status}
\alias{mitm_status}
\title{Check on the status of an mitmproxy process created with \code{mitm_start()}}
\usage{
mitm_status(pid_obj)
}
\arguments{
\item{#'}{@param pid_obj \code{mitm_pid} object created with \code{start_mitm}}
}
\value{
logical. \code{TRUE} if the background process is still running
}
\description{
Check on the status of an mitmproxy process created with \code{mitm_start()}
}
