% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-httr.R
\name{read_httr}
\alias{read_httr}
\title{Read a serialed httr ndjson object file}
\usage{
read_httr(path)
}
\arguments{
\item{path}{path to httr ndjson object file}
}
\description{
Read a serialed httr ndjson object file
}
