% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitm-helpers.R
\name{start_mitm}
\alias{start_mitm}
\title{Spawn a mitmdump background process in "HAR capture" mode and return the process id}
\usage{
start_mitm(extra_args = NULL)
}
\arguments{
\item{extra_args}{see \code{args} in \code{\link[sys:exec_background]{sys::exec_background()}}}
}
\value{
\code{mitm_pid} object
}
\description{
Returns \code{mitm_pid} object which can be used with \code{\link[=stop_mitm]{stop_mitm()}} to kill
the mitmproxy process and retrieve the generated HAR file or \code{\link[=mitm_status]{mitm_status()}}
to check on the status of the background process.
}
\seealso{
Other mitm_helpers: \code{\link{call_mitm}},
  \code{\link{install_mitm}}, \code{\link{mitm_help}}
}
\concept{mitm_helpers}
