% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitm-helpers.R
\name{stop_mitm}
\alias{stop_mitm}
\title{Stop the mitmproxy background process and retrieve the generated HAR file}
\usage{
stop_mitm(pid_obj, read = TRUE, save = NULL)
}
\arguments{
\item{pid_obj}{\code{mitm_pid} object created with \code{start_mitm}}

\item{read}{if \code{TRUE} (the default) read in the \code{httr} or HAR file. NOTE: you are
encouraged to set this to \code{FALSE} if you're using HAR format and
reading a large number of  pages or pages with heavy resources. \code{\link[HARtools:readHAR]{HARtools::readHAR()}}}

\item{save}{if not \code{NULL} then the \code{httr} or HAR file generated by \code{mitmproxy} will be
sent to this filesystem location (\strong{you need to use a path that includes a
filename}). This will be \code{path.expand()}ed. It is recommended that you use
\code{.json.gz} as the file extension for the \code{httr} object type.}
}
\value{
a \code{HARtools} object if \code{read} was \code{TRUE} and the collection was \code{har};
a list of (minimal) \code{httr} \code{response} objects; nothing if \code{read} was \code{FALSE}.
}
\description{
Stop the mitmproxy background process and retrieve the generated HAR file
}
