% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitm-helpers.R
\name{call_mitm}
\alias{call_mitm}
\title{Spawn a mitmdump background process with custom args and return the process id}
\usage{
call_mitm(args)
}
\arguments{
\item{args}{see \code{\link[sys:exec_background]{sys::exec_background()}}}
}
\description{
Returns process id. Use \code{\link[tools:pskill]{tools::pskill()}} to kill.
}
\seealso{
Other mitm_helpers: \code{\link{install_mitm}},
  \code{\link{mitm_help}}, \code{\link{start_mitm}}
}
