% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitm.R, R/zzz.R
\name{mitm_check}
\alias{mitm_check}
\alias{mitm_check}
\title{MITM check}
\usage{
mitm_check(timeout = 4)

mitm_check(timeout = 4)
}
\arguments{
\item{timeout}{Wait time after launching MITM server to run R code}

\item{package_path}{Path to a local copy of a R package}

\item{timeout}{}
}
\value{
list

list
}
\description{
MITM check

MITM check
}
\examples{
\dontrun{
mitm_check(package_path = "/local/path")
}
mitm_check()
}
