% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-httr.R
\name{read_httr}
\alias{read_httr}
\title{Read a serialized httr response object ndjson file}
\usage{
read_httr(path)
}
\arguments{
\item{path}{path to httr ndjson object file. Can be \code{.json} or \code{.json.gz}}
}
\description{
Read a serialized httr response object ndjson file
}
