% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitm-helpers.R
\name{stop_mitm}
\alias{stop_mitm}
\title{Stop the mitmproxy background process and retrieve the generated HAR file}
\usage{
stop_mitm(pid_obj)
}
\arguments{
\item{pid_obj}{\code{mitm_pid} object created with \code{start_mitm}}
}
\value{
\code{HARtools} object or \code{NULL}
}
\description{
Stop the mitmproxy background process and retrieve the generated HAR file
}
