% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitm-helpers.R
\name{stop_mitm}
\alias{stop_mitm}
\title{Stop the mitmproxy background process and retrieve the generated HAR file}
\usage{
stop_mitm(pid_obj, read_har = TRUE, save_har = NULL)
}
\arguments{
\item{pid_obj}{\code{mitm_pid} object created with \code{start_mitm}}

\item{read_har}{if \code{TRUE} (the default) read in the HAR file. NOTE: you are
encouraged to set this to \code{FALSE} if you're reading a large number of
pages or pages with heavy resources. \code{\link[HARtools:readHAR]{HARtools::readHAR()}} does not handle
very large HAR files well.}

\item{save_har}{if not \code{NULL} then the HAR file generated by \code{mitmproxy} will be
sent to this filesystem location (\strong{you need to use a path that includes a
filename}). This will be \code{path.expand()}ed.}
}
\value{
a \code{HARtools} object if \code{read_har} was \code{TRUE}.
}
\description{
Stop the mitmproxy background process and retrieve the generated HAR file
}
